% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Construct.R, R/package-RVenn.R
\docType{methods}
\name{construct,ANY-method}
\alias{construct,ANY-method}
\alias{construct}
\title{Build a \code{Venn} object}
\usage{
\S4method{construct}{ANY}(sets)
}
\arguments{
\item{sets}{(Required) A \code{list} containing vectors in the same type. If
a vector contains duplicates, they will be discarded. If the \code{list}
doesn't have names, the sets will be named as "Set_1", "Set_2", "Set_3" and
so on.}
}
\value{
A \code{Venn} object.
}
\description{
\code{construct} builds a \code{Venn} object from a \code{list}.
}
\examples{
venn = construct(list(letters[1:10], letters[3:12], letters[6:15]))
print(venn)
}
