% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-RVenn.R
\docType{package}
\name{RVenn}
\alias{RVenn}
\alias{RVenn-package}
\title{\code{RVenn}: A package for set operations for three or more sets.}
\description{
The \code{RVenn} package provides functions for application of set operations on
multiple sets. It also performs hypergeometric test and plots the Venn
diagram for 2-3 sets by using \code{ggplot2} and \code{ggforce} packages.
}
