% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Construct.R, R/RVenn.R
\docType{methods}
\name{Venn,ANY-method}
\alias{Venn,ANY-method}
\alias{Venn}
\title{Build a \code{Venn} object.}
\usage{
\S4method{Venn}{ANY}(sets)
}
\arguments{
\item{sets}{(Required) A list containing vectors in the same class. If a
vector contains duplicates they will be discarded. If the list doesn't have
names the sets will be named as "Set_1", "Set_2", "Set_3" and so on.}
}
\value{
A \code{Venn} object.
}
\description{
\code{Venn} builds a \code{Venn} object from a list.
}
\examples{
venn = Venn(list(letters[1:10], letters[3:12], letters[6:15]))
print(venn)
}
