% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/venn2.R
\name{venn2}
\alias{venn2}
\title{Draw pairwise venn diagram}
\usage{
venn2(a, b, ab, name = c("", ""), border = "black", col = c(rgb(1, 0, 0,
  0.2), rgb(0, 0, 1, 0.2)), main = "", padding = c(0, 0.15, 0.3, 0.15),
  cex = 1)
}
\arguments{
\item{a}{size of set A}

\item{b}{size of set B}

\item{ab}{size of shared part between A & B}

\item{name}{names of A & B}

\item{border}{color of circle's border}

\item{col}{colors of A & B}

\item{main}{main title of the diagram}

\item{padding}{padding size of the diagram, for display names}

\item{cex}{change font size}
}
\description{
Draw pairwise venn diagram
}
\examples{
venn2(7, 10, 3)
}

