% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVenn.R
\docType{package}
\name{RVenn}
\alias{RVenn}
\alias{RVenn-package}
\title{\code{RVenn}: A package for set operations for many sets.}
\description{
Set operations for many sets. The base functions for set operations in R can
be used for only two sets. This package uses 'purr' to find the union,
intersection and difference of three or more sets. This package also provides
functions for pairwise set operations among several sets. Further, based on
'ggplot2' and 'ggforce', a Venn diagram can be drawn for two or three sets.
For bigger data sets, a clustered heatmap showing presence/absence of the
elements of the sets can be drawn based on the 'pheatmap' package. Finally,
enrichment test can be applied to two sets whether an overlap is
statistically significant or not.
}
